Moved services to folders depending on the type. #82
