+++ src/app/shared/contracts/kudos-organisations.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed, inject } from '@angular/core/testing';
 
-import { Web3Service } from './web3.service';
-import { KudosPollFactoryService } from './kudos-poll-factory.service';
+import { Web3Service } from '../web3.service';
+import { KudosPollFactoryService } from '../kudos-poll-factory.service';
 
 import { KudosOrganisationsService } from './kudos-organisations.service';
