+++ src/app/shared/contracts/kudos-organisations.service.ts
@@ -4,11 +4,11 @@ import 'rxjs/add/operator/filter';
 import 'rxjs/add/operator/first';
 import 'rxjs/add/operator/toPromise';
 
-import * as KudosOrganisationsDefinition from '../../../build/contracts/KudosOrganisations.json';
+import * as KudosOrganisationsDefinition from '../../../../build/contracts/KudosOrganisations.json';
 
 import { SmartContract } from './smart-contract.abstract';
 import { Contract } from './truffle.interface';
-import { Web3Service, ConnectionStatus } from './web3.service';
+import { Web3Service, ConnectionStatus } from '../web3.service';
 
 interface KudosOrganisationsConstants {
   getOrganisations: string[];
