+++ src/app/shared/contracts/kudos-poll.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed, inject } from '@angular/core/testing';
 
-import { Web3Service } from './web3.service';
+import { Web3Service } from '../web3.service';
 
 import { KudosPollService } from './kudos-poll.service';
