+++ src/app/shared/contracts/truffle.interface.ts
