+++ src/app/shared/guards/is-owner.guard.ts
@@ -4,29 +4,26 @@ import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/fromPromise';
 import 'rxjs/add/operator/do';
 import 'rxjs/add/operator/first';
-import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/mergeMap';
 
-import { KudosTokenFactoryService } from './kudos-token-factory.service';
+import { KudosTokenFactoryService } from '../kudos-token-factory.service';
 
 @Injectable()
-export class IsPollGuard implements CanActivate {
+export class IsOwnerGuard implements CanActivate {
 
   constructor(private kudosTokenFactoryService: KudosTokenFactoryService, private router: Router) { }
 
   canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
     const tokenAddress = next.params.tokenAddress || next.parent.params.tokenAddress;
-    const pollAddress = next.params.address;
     const kudosTokenService = this.kudosTokenFactoryService
       .getKudosTokenServiceAt(tokenAddress);
     return kudosTokenService
       .onInitialized
       .first()
-      .mergeMap(() => Observable.fromPromise(kudosTokenService.getPreviousPolls()))
-      .map(polls => polls.indexOf(pollAddress) !== -1)
+      .mergeMap(() => Observable.fromPromise(kudosTokenService.imOwner()))
       .do(imOwner => {
         if (!imOwner) {
-          this.router.navigate([state.url.split('/').slice(0, -2).join('/')]);
+          this.router.navigate([state.url.split('/').slice(0, -1).join('/')]);
         }
       });
   }
