+++ src/app/shared/guards/is-poll.guard.spec.ts
@@ -1,10 +1,10 @@
 import { TestBed, async, inject } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 
-import { Web3Service } from './web3.service';
-import { KudosTokenService } from './kudos-token.service';
-import { KudosTokenFactoryService } from './kudos-token-factory.service';
-import { KudosPollFactoryService } from './kudos-poll-factory.service';
+import { Web3Service } from '../web3.service';
+import { KudosTokenService } from '../contracts/kudos-token.service';
+import { KudosTokenFactoryService } from '../kudos-token-factory.service';
+import { KudosPollFactoryService } from '../kudos-poll-factory.service';
 
 import { IsPollGuard } from './is-poll.guard';
