+++ src/app/shared/guards/is-token.guard.spec.ts
@@ -1,20 +1,20 @@
 import { TestBed, async, inject } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 
-import { Web3Service } from './web3.service';
-import { KudosTokenFactoryService } from './kudos-token-factory.service';
-import { KudosPollFactoryService } from './kudos-poll-factory.service';
+import { Web3Service } from '../web3.service';
+import { KudosTokenFactoryService } from '../kudos-token-factory.service';
+import { KudosPollFactoryService } from '../kudos-poll-factory.service';
 
-import { IsOwnerGuard } from './is-owner.guard';
+import { IsTokenGuard } from './is-token.guard';
 
-describe('IsOwnerGuard', () => {
+describe('IsTokenGuard', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
         RouterTestingModule,
       ],
       providers: [
-        IsOwnerGuard,
+        IsTokenGuard,
         Web3Service,
         KudosTokenFactoryService,
         KudosPollFactoryService,
@@ -22,7 +22,7 @@ describe('IsOwnerGuard', () => {
     });
   });
 
-  it('should create', inject([IsOwnerGuard], (guard: IsOwnerGuard) => {
+  it('should create', inject([IsTokenGuard], (guard: IsTokenGuard) => {
     expect(guard).toBeTruthy();
   }));
 });
